Update README.md
Using the given example throws an error about missing a name of the device. As I am unfamiliar with the code and with PACE5000 I struggled the whole day to figure the error out and just adding the name in the config file fixed everything.
So, I am proposing to change the example for other people trying out this library.
P.S. Thanks for this amazing project!